+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 5.6.0 (2024-02-14)
+### Features Added
+
+- New enum type `ZonePlacementPolicyType` with values `ZonePlacementPolicyTypeAny`
+- New struct `Placement`
+- New field `Placement` in struct `VirtualMachine`
+
+
 ## 5.5.0 (2024-01-26)
 ### Features Added